+++ src/brom_drake/all/__init__.py
@@ -5,7 +5,7 @@ from brom_drake.PortWatcher.port_watcher import PortWatcher
 from brom_drake.PortWatcher.port_watcher_options import PortWatcherOptions, FigureNamingConvention
 from brom_drake.PortWatcher.port_figure_arrangement import PortFigureArrangement
 from brom_drake.example_helpers import BlockHandlerSystem
-from brom_drake.urdf import DrakeReadyURDFConverter, drakeify_my_urdf
+from brom_drake.file_manipulation.urdf import DrakeReadyURDFConverter, drakeify_my_urdf
 
 __all__ = [
     'DiagramTarget', 'DiagramWatcher',
