+++ src/brom_drake/file_manipulation/urdf/simple_shape_urdfs/shape_definition.py
