+++ src/brom_drake/file_manipulation/urdf/simple_shape_urdfs/urdf_definition.py
